+++ test/td_dd/grants/requests_test.exs
@@ -41,8 +41,9 @@ defmodule TdDd.Grants.RequestsTest do
     end
 
     test "create_grant_request_group/2 with valid data creates a grant_request_group" do
-      %{id: data_structure_id} = insert(:data_structure)
-      %{user_id: user_id} = build(:claims)
+      %{id: domain_id} = CacheHelpers.insert_domain()
+      %{id: data_structure_id} = insert(:data_structure, domain_ids: [domain_id])
+      %{user_id: user_id} = claims = build(:claims)
 
       params = %{
         type: @template_name,
@@ -52,7 +53,7 @@ defmodule TdDd.Grants.RequestsTest do
       }
 
       assert {:ok, %{group: group, statuses: statuses, requests: {_count, [request_id]}}} =
-               Requests.create_grant_request_group(params)
+               Requests.create_grant_request_group(params, claims)
 
       assert %{
                type: @template_name,
@@ -65,8 +66,9 @@ defmodule TdDd.Grants.RequestsTest do
     end
 
     test "create_grant_request_group/2 with modification_grant" do
-      %{id: data_structure_id} = insert(:data_structure)
-      %{user_id: user_id} = build(:claims)
+      %{id: domain_id} = CacheHelpers.insert_domain()
+      %{id: data_structure_id} = insert(:data_structure, domain_ids: [domain_id])
+      %{user_id: user_id} = claims = build(:claims)
       %{id: grant_id} = modification_grant = insert(:grant, data_structure_id: data_structure_id)
 
       params = %{
@@ -77,7 +79,7 @@ defmodule TdDd.Grants.RequestsTest do
       }
 
       assert {:ok, %{group: group}} =
-               Requests.create_grant_request_group(params, modification_grant)
+               Requests.create_grant_request_group(params, claims, modification_grant)
 
       assert %{
                type: @template_name,
@@ -87,9 +89,11 @@ defmodule TdDd.Grants.RequestsTest do
     end
 
     test "creates grant_request_group requests" do
-      %{id: ds_id_1} = insert(:data_structure)
-      %{id: ds_id_2} = insert(:data_structure)
-      %{user_id: user_id} = build(:claims)
+      %{id: domain_id} = CacheHelpers.insert_domain()
+      domain_ids = [domain_id]
+      %{id: ds_id_1} = insert(:data_structure, domain_ids: domain_ids)
+      %{id: ds_id_2} = insert(:data_structure, domain_ids: domain_ids)
+      %{user_id: user_id} = claims = build(:claims)
 
       requests = [
         %{
@@ -107,7 +111,8 @@ defmodule TdDd.Grants.RequestsTest do
         created_by_id: user_id
       }
 
-      assert {:ok, %{group: %{requests: requests}}} = Requests.create_grant_request_group(params)
+      assert {:ok, %{group: %{requests: requests}}} =
+               Requests.create_grant_request_group(params, claims)
 
       assert [
                %{